Tweaked itoa and uitoa for long numbers

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: aperez-b <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/05 17:32:34 by aperez-b          #+#    #+#             */
-/*   Updated: 2021/05/01 10:46:28 by aperez-b         ###   ########.fr       */
+/*   Updated: 2021/05/11 13:39:42 by aperez-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,8 +31,8 @@ size_t	ft_strlcat(char *dst, const char *src, size_t size);
 size_t	ft_strlcpy(char *dst, const char *src, size_t size);
 int		ft_strncmp(const char *s1, const char *s2, size_t n);
 int		ft_atoi(const char *nptr);
-char	*ft_itoa(int n);
-char	*ft_uitoa(unsigned int n);
+char	*ft_itoa(long n);
+char	*ft_uitoa(long long unsigned n);
 int		ft_putchar_fd(char c, int fd);
 int		ft_putnchar_fd(char c, int fd, int n);
 int		ft_putendl_fd(char *s, int fd);
